Modified tests and Add more tests
