+++ package.json
@@ -12,7 +12,7 @@
     "migrate": "npm run unmigrate && node ./node_modules/sequelize-cli/lib/sequelize db:migrate && npm run seed",
     "heroku-postbuild": "npm run migrate",
     "postinstall": "npm run build",
-    "test": "nyc --reporter=html --reporter=text mocha server/tests/ --exit --compilers js:babel-register",
+    "test": "npm run migrate && nyc --reporter=html --reporter=text mocha server/tests/ --exit --compilers js:babel-register",
     "test-watch": "nodemon --exec \"npm test\"",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "coveralls": "nyc --reporter=lcov --reporter=text-lcov npm test",
